+++ apps/server/src/modules/email/controllers/email.controller.ts
@@ -4,7 +4,7 @@ import { ApiOperation } from '@nestjs/swagger';
 
 import { ChartService } from '@/modules/chart/services/chart.service';
 import { ProjectService } from '@/modules/project/services/project.service';
-import { formatDate, lastMonthRange, lastWeekRange } from '@/utils';
+import { lastMonthRange, lastWeekRange } from '@/utils';
 import { EmailService } from '../services/email.service';
 
 @Controller('email')
@@ -36,18 +36,20 @@ export class EmailController {
                     try {
                         await this.sendProject({ projectId, emails });
                     } catch (error) {}
-                    console.log(formatDate(), ` ${name}, 定时发送单个子产品的数据周报到指定邮箱`);
+                    console.log(
+                        `\nprojectId: ${projectId}, ${name}, 发送【单个子产品】的数据周报到指定邮箱`
+                    );
                 }
             });
     }
     async handleSendAll() {
         if (process.env.DEFAULT_EMAIL) {
-            this.sendAll({ emails: process.env.DEFAULT_EMAIL });
-            console.log(formatDate(), ' 定时发送所有子产品的数据周报到指定邮箱');
+            console.log('\n发送【所有子产品】的数据周报到指定邮箱');
+            await this.sendAll({ emails: process.env.DEFAULT_EMAIL });
         }
     }
 
-    @ApiOperation({ summary: '发送邮件' })
+    @ApiOperation({ summary: '发送【单个子产品】的数据周报到指定邮箱' })
     @HttpCode(HttpStatus.OK)
     @Post('sendProjectMail')
     async sendProject(@Body() { projectId, emails: _emails }) {
@@ -56,10 +58,10 @@ export class EmailController {
 
         if (emails?.split(',')?.length) {
             try {
-                const result = await this.generatePromise(project);
+                const result = await this.generatePromise({ ...project, emails });
                 return result;
             } catch (error) {
-                console.log(formatDate(), ' 邮件数据处理失败', error);
+                console.log(`\nprojectId: ${projectId}, ${project?.name}, 邮件数据处理失败`, error);
                 throw new HttpException('没有历史检测数据', HttpStatus.OK);
             }
         }
@@ -82,7 +84,6 @@ export class EmailController {
         if (projectChartData?.versionNameList?.length) {
             const result = await this.emailService.sendMail(
                 project,
-                lastWeekRange,
                 projectChartData,
                 fileSizeChartData
             );
@@ -92,7 +93,7 @@ export class EmailController {
         }
     }
 
-    @ApiOperation({ summary: '定时发送所有子产品的数据周报到指定邮箱' })
+    @ApiOperation({ summary: '发送【所有子产品】的数据周报到指定邮箱' })
     @HttpCode(HttpStatus.OK)
     @Post('sendAll')
     async sendAll(@Body() { emails = process.env.DEFAULT_EMAIL }) {
@@ -102,40 +103,50 @@ export class EmailController {
             }
 
             const [startTime, endTime] = lastWeekRange;
+            const [startMonthTime, endMonthTime] = lastMonthRange;
             let projectList = await this.projectService.findAll();
             projectList = projectList.filter((project) => project.name !== '汇总');
 
-            const promiseList = projectList.map((project) => {
+            const projectChartDataPromiseList = projectList.map((project) => {
                 return this.chartService.projectChart({
                     projectId: project.projectId,
                     startTime,
                     endTime,
                 });
             });
+            const fileSizeChartPromiseList = projectList.map((project) => {
+                return this.chartService.fileSizeChart({
+                    projectId: project.projectId,
+                    startTime: startMonthTime,
+                    endTime: endMonthTime,
+                });
+            });
 
-            const projectChartDataList = [];
-            const projectChartDataResults = await Promise.all(promiseList);
+            const chartDataList = [];
+            const projectChartDataResults = await Promise.all(projectChartDataPromiseList);
+            const fileSizeChartDataResults = await Promise.all(fileSizeChartPromiseList);
+
             for (let i = 0; i < projectChartDataResults.length; i++) {
                 if (projectChartDataResults[i].versionNameList?.length) {
                     const { projectId, name } = projectList[i];
-                    projectChartDataList.push({
+                    chartDataList.push({
                         projectId,
                         name,
                         projectChartData: projectChartDataResults[i],
+                        fileSizeChartData: fileSizeChartDataResults[i],
                     });
                 }
             }
 
-            if (projectChartDataList.length) {
-                const result = await this.emailService.sendMailAllProject(
-                    emails,
-                    projectChartDataList,
-                    lastWeekRange
-                );
+            if (chartDataList.length) {
+                const result = await this.emailService.sendMailAllProject(emails, chartDataList);
                 return result;
             }
         } catch (error) {
-            throw new HttpException(`尝试发送所有子产品的数据周报失败, ${error}`, HttpStatus.OK);
+            throw new HttpException(
+                `尝试发送【所有子产品】的数据周报失败, ${error}`,
+                HttpStatus.OK
+            );
         }
     }
 }