+++ apps/server/src/modules/email/services/email.service.ts
@@ -2,7 +2,7 @@ import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { MailerService } from '@nestjs-modules/mailer';
 import { IFileSizeChartData, IProjectChartData } from 'typing';
 
-import { formatDate } from '@/utils';
+import { lastWeekRange } from '@/utils';
 import { renderChart } from '@/utils/echarts';
 
 @Injectable()
@@ -11,7 +11,6 @@ export class EmailService {
 
     async sendMail(
         project,
-        lastWeekRange,
         projectChartData: IProjectChartData,
         fileSizeChartData: IFileSizeChartData
     ) {
@@ -26,32 +25,39 @@ export class EmailService {
                     subject: `【${name}】易测数据周报（${startTime}~${endTime}）`, // 标题
                     html,
                 });
-                console.log(formatDate(), ' 发送单个子产品的数据周报成功', result);
+                console.log(
+                    `\nprojectId: ${projectId}, ${name}, 发送【单个子产品】的数据周报成功`,
+                    result
+                );
                 return result;
             } catch (error) {
-                console.log(formatDate(), ' 尝试发送单个子产品的数据周报失败', error);
+                console.log(
+                    `\nprojectId: ${projectId}, ${name}, 尝试发送【单个子产品】的数据周报失败`,
+                    error
+                );
                 throw new HttpException('尝试发送单个子产品的数据周报失败', HttpStatus.OK);
             }
         }
     }
 
     // 定时发送所有子产品的数据周报到指定邮箱
-    async sendMailAllProject(emails, projectChartDataList = [], lastWeekRange) {
+    async sendMailAllProject(emails, chartDataList = []) {
         const [_, __, startTime, endTime] = lastWeekRange;
 
         if (emails?.split(',')?.length) {
-            const html = renderChart(projectChartDataList);
+            const html = renderChart(chartDataList);
             try {
                 const result = await this.mailerService.sendMail({
                     to: emails?.split(','),
                     subject: `【数栈子产品】易测数据周报（${startTime}~${endTime}）`, // 标题
                     html,
                 });
-                console.log(formatDate(), ' 发送所有子产品的数据周报成功', result);
+
+                console.log('\n发送【所有子产品】的数据周报成功', result);
                 return result;
             } catch (error) {
-                console.log(formatDate(), ' 尝试发送所有子产品的数据周报失败', error);
-                throw new HttpException('尝试发送所有子产品的数据周报失败', HttpStatus.OK);
+                console.log('\n尝试发送【所有子产品】的数据周报失败', error);
+                throw new HttpException('尝试发送【所有子产品】的数据周报失败', HttpStatus.OK);
             }
         }
     }