+++ scripts/start.sh
@@ -1,3 +1,3 @@
 # pnpm -v
 pnpm install --production
-pnpm pm2
+pm2 restart ./pm2/config.json --env production