+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.1.2](///compare/v2.1.1...v2.1.2) (2024-07-16)
+
+
+### Features
+
+* log optimize 6fbdd17
+
+### [2.1.1](///compare/v2.1.0...v2.1.1) (2024-07-16)
+
+
+### Features
+
+* console log optimize b5d1edc
+* week data email optimize 373488b
+
+
+### Bug Fixes
+
+* send email when email not saved d9e8e17
+
+## [2.1.0](///compare/v2.0.2...v2.1.0) (2024-07-15)
+
+
+### Features
+
+* log optimize c182fcb
+
+
+### Bug Fixes
+
+* sendAll need fileSizeChartData 1eb96df
+
 ### [2.0.2](///compare/v2.0.1...v2.0.2) (2024-07-12)